[REFACTOR: bridge] 클래스의 역할에 따라 controller, domain, utils, view로 package 분류
