+++ src/main/java/bridge/domain/BridgeMaker.java
@@ -1,4 +1,6 @@
-package bridge;
+package bridge.domain;
+
+import bridge.utils.BridgeNumberGenerator;
 
 import java.util.ArrayList;
 import java.util.List;
