+++ src/main/java/bridge/utils/BridgeRandomNumberGenerator.java
@@ -1,5 +1,6 @@
-package bridge;
+package bridge.utils;
 
+import bridge.utils.BridgeNumberGenerator;
 import camp.nextstep.edu.missionutils.Randoms;
 
 public class BridgeRandomNumberGenerator implements BridgeNumberGenerator {
